requirements.txt
absl-py==0.9.0
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
grpcio==1.31.0
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
mypy==0.782
mypy-extensions==0.4.3
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pep8==1.7.1
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
pytorch-lightning==0.9.0rc15
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
toml==0.10.1
torch==1.6.0
torchsummary==1.5.1
torchvision==0.7.0
tqdm==4.48.2
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
